requirements.txt
streamlit
transformers
torch
sentence-transformers
faiss-cpu
pypdf
numpy
scikit-learn
rouge-score
psutil